chore: go vs toolchain directive
When a `toolchain` directive is provided in go.mod, it takes precedence
over the `go` directive to establish the go version when running a go
command. This was revealed by #41, when cometbft was upgraded.

This creates issue with our go version enforcement that relies only on
the `go` directive. However it is possible to ignore the `toolchain`
directive by setting the `GOTOOLCHAIN=local` env, this is what this change
is about.

TODO: Maybe this should be set in Makefile too.